Implement V4.1 CED and optimize serving hot paths
Prepare full encoder and decoder global KV before bounded decoder replay. Preserve prompt-logprob row contracts, request-private decoder SWA, prefix-reuse boundaries, DSpark context selection, and CUDA graph capture safety.

Use whole-batch native attention, bounded indexer visibility, FP32 routed expert composition and dtype-bound PCIe collectives. Stream packed Engram shards into mapped host allocations, but keep SSD as the serving default after full-RAM host OOM.

Interim checkpoint requested without further validation. Focused CED/native-attention integration previously passed 28 tests. Commit hooks applied formatting but reported unresolved lint, typing, forbidden-import and torch.cuda API checks; hooks bypassed for this checkpoint.

+++ vllm/models/deepseek_v4_1/nvidia/b12x_moe.py
@@ -144,7 +144,7 @@ class B12xV41Experts(DeepseekV4MegaMoEExperts):
         fused_moe.run(binding=binding)
 
     def forward(self, x, weights, ids):
-        out = torch.empty_like(x)
+        out = torch.empty_like(x, dtype=torch.float32)
         _experts(x, weights, ids, out, self.prefix)
         return out
